### 2017-10-25 #439 @whedon
[PRE REVIEW]: mpnum: A matrix product representation library for Python
**Submitting author:** @dseuss (<a href="http://orcid.org/0000-0002-6354-457X">Daniel Suess</a>)
**Repository:** <a href="https://github.com/dseuss/mpnum" target ="_blank">https://github.com/dseuss/mpnum</a>
**Version:** v1.0.0
