### 2017-12-10 #465 @labarba
Hi @SylvainCorlay — ping! Gentle reminder here about the review for JOSS. You can start on it, and report any issues you encounter right here (or post issues on the repo of the submitted software). Cheers!
