### 2017-12-10 #465 @dseuss
After the submission, we noted a compatibility bug in the `setup.py` that prevents v1.0.0 to be installed with Python2. Could we bump the refereed version to v1.0.1.
