### 2017-12-11 #465 @SylvainCorlay
> Does that mean 1.0.1 doesn't work for you either?

Absolutely! Pardon my brevity.
