### 2017-12-11 #465 @SylvainCorlay
Note: I checked the box regarding the matched version. I don't know if it matters that the generated issue specifies 1.0.0
