### 2017-12-11 #465 @SylvainCorlay
As I am going through the checklist:

>  References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

I found that there is a DOI for the reference **Enabling High-Dimensional Hierarchical Uncertainty Quantification by ANOVA and Tensor-Train Decomposition*.
