### 2017-12-11 #465 @dseuss
@SylvainCorlay Thanks for the feedback.

> I found that there is a DOI for the reference Enabling High-Dimensional Hierarchical Uncertainty Quantification by ANOVA and Tensor-Train Decomposition.
Added DOI and updated arXiv references for another paper  in `a3aaaf5`

> Very minor: on this point, a "contributing" section in the readme could be added (simply linking to the relevant section of the documentation).
Done.

> For some reason, I had to install hdf5 to be able to run the tests. The test_requires specify h5py, which probably requires some installation of hdf5. It seems that the test requirements won't use wheels.
Right, I recalled that we had to manually install test dependencies on travis for that reason. I have added a note to the README as well as to the [documentation](http://mpnum.readthedocs.io/en/latest/devel.html). (The latter should appear once readthedocs renders the new version)
