### 2017-12-11 #465 @SylvainCorlay
I could install, run the tests, and the example notebook. The whole process was straightforward. The need for the package is properly justified and prior work appropriately cited. Documentation includes both high-level narrative documentation introducing the package and an API reference generated from the doc strings (although a few items are left as todo).

1) Quick note about the example notebook: I find it extremely useful and detailed. A link to the notebook is provided in the documentation, but the rendered notebook itself could probably be included *as-is* in the documentation. A good tool to achieve this is to use the [nbsphinx](https://nbsphinx.readthedocs.io/en/0.2.18/) sphinx extension.

Minor suggestion: the notebook could include some visual from the narrative documentation about the graphical notation, and how it translate with a simple code example.

3) About the narrative documentation:  *the matrix product state representation of a state* comes without context. The rest of the paragraph is more self-explanatory but I think that this is domain-specific, and is probably not necessary for introducing MPS and MPO.
