### 2017-12-12 #465 @labarba
@SylvainCorlay : I see that all boxes are checked. Can you confirm that you recommend acceptance now?
