### 2017-12-12 #465 @dseuss
Thanks a lot @SylvainCorlay for the helpful comments. We already implemented remark 1 on the master branch in the repo. We will address the other remarks shortly after fixing some slight mistakes in some images.

@labarba I would move the `v1.0.1` tag to a different commit in the repo once we incorporate all the fixes. Since all the fixes are in the documentation, this does not affect the version on PyPi. Would that be okay with you?
