### 2017-12-12 #465 @labarba
Yay! @dseuss —We'll now need you to deposit the final release on an archival host like figshare or Zenodo, and report here the DOI. Then @arfon will press the buttons to publish!
