### 2017-12-13 #465 @dseuss
I have uploaded the final version v1.0.2 with the DOI [10.5281/zenodo.1115574](https://doi.org/10.5281/zenodo.1115574). We have updated the refereed version according to the suggestions of the referee:

- Added some images to the notebook
- Embedded the notebook on readthedocs

We also migrated all the figures from Inkscape to LaTeX/TikZ and fixed some notation errors.
