### 2017-12-14 #465 @dseuss
Could we wait for another hour? I just noticed that the bibliography in the PDF is everything but consistent. I'll try to fix this now.
