### 2017-12-14 #465 @dseuss
OK, now I am happy with it as well. Just cleaned up the names in the references, removed unnecessary fields from the bib file, and fixed citations of the Oseledets paper, which gave ??? in the old version.
