### 2017-12-15 d5c2342
Improved the code
according to the discussion on the last commit.
