### 2017-12-15 #465 @milan-hl
I slightly adapted the spelling of one of the references. I'm looking forward to the paper being published. Thanks everyone!
