### 2017-12-15 c78ece8
Improved the code
according to the discussion on the last commit.
