### 2017-12-15 #465 @arfon
@dseuss - your submission is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00465 ⚡️ 🚀 💥
